restructured component hierarchy
