+++ src/components/Contacts.js
@@ -1,0 +1,9 @@
+import * as React from 'react';
+
+function Contacts() {
+    return (
+        <div></div>
+    )
+}
+
+export default Contacts
