+++ src/components/Home.js
@@ -1,0 +1,10 @@
+import * as React from "react";
+
+
+function Home() {
+    return (
+        <div></div>
+    )
+}
+
+export default Home;
