+++ src/components/Profile.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function Profile() {
+    return (
+        <div></div>
+    )
+}
+
+export default Profile;
